Package Development

+++ Website/htdocs/mpmanager/database/migrations/2020_10_02_072509_create_time_of_usages_table.php
@@ -18,7 +18,7 @@ class CreateTimeOfUsagesTable extends Migration
             $table->unsignedInteger('tariff_id');
             $table->string('start');
             $table->string('end');
-            $table->integer('value');
+            $table->double('value');
             $table->timestamps();
         });
     }
